+++ lib/solid_stack_web.rb
@@ -6,7 +6,7 @@ module SolidStackWeb
     attr_writer :page_size, :connects_to, :slow_job_threshold,
                 :alert_webhook_url, :alert_webhook_cooldown,
                 :alert_failure_threshold, :alert_queue_thresholds,
-                :alert_slow_job_count_threshold,
+                :alert_slow_job_count_threshold, :alert_stale_process_threshold,
                 :dashboard_refresh_interval, :default_refresh_interval,
                 :search_results_limit, :allow_value_preview
 
@@ -40,6 +40,10 @@ module SolidStackWeb
 
     def alert_slow_job_count_threshold
       @alert_slow_job_count_threshold
+    end
+
+    def alert_stale_process_threshold
+      @alert_stale_process_threshold
     end
 
     def dashboard_refresh_interval